% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyMEP_dose_response_curves.R
\name{asymptotic_prediction}
\alias{asymptotic_prediction}
\title{DRC function: asymptotic_prediction}
\usage{
asymptotic_prediction(x, params)
}
\arguments{
\item{x}{input variable}

\item{params}{list of input parameter;
Asym  a numeric parameter representing the horizontal asymptote
on the right side (very large values of input). ;
lrc     a numeric parameter representing the natural
 logarithm of the rate constant;
 c0      a numeric parameter representing the x for which the response is zero.}
}
\description{
DRC function: asymptotic_prediction
}
\examples{
asymptotic_prediction(5,list("Asym_value"=0.5,
                    "lrc_value"=0.2,
                      "c0_value"=4))
# visualization
asymptote <- lapply(seq(0, 10, 0.1),
asymptotic_prediction,
list("Asym_value"=0.5, "lrc_value"=0.2, "c0_value"=4))
plot(seq(0, 10, 0.1), asymptote)
}
\keyword{DRC}
