% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyMEP_prediction_helper_functions.R
\name{available_crops_and_phases}
\alias{available_crops_and_phases}
\title{available_crops_and_phases}
\usage{
available_crops_and_phases()
}
\value{
dataframe with available crops and phenology phases
}
\description{
check what crops and corresponding phenology phases are available at the
 moment. Chose a crop (crop_abbrev), phenological phase from the output of
 this function further usage
}
\examples{
available_crops_and_phases()
}
\keyword{and}
\keyword{are}
\keyword{available}
\keyword{check}
\keyword{corresponding}
\keyword{crops}
\keyword{phases}
\keyword{phenology}
\keyword{what}
