% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyMEP_prediction_helper_functions.R
\name{available_environmental_covariates}
\alias{available_environmental_covariates}
\title{available_environmental_covariates
check what environmental covariates are implemented, use or alter prediction
  these abbreviations and the corresponding unit}
\usage{
available_environmental_covariates()
}
\description{
available_environmental_covariates
check what environmental covariates are implemented, use or alter prediction
  these abbreviations and the corresponding unit
}
\examples{
available_environmental_covariates()
}
\keyword{are}
\keyword{check}
\keyword{covariates}
\keyword{environmental}
\keyword{implemented}
\keyword{what}
