% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/DyMEP_apply_unknown_environment_prediction_utils.R
\name{envpredutils.cumulative_dose_response_pred_helper}
\alias{envpredutils.cumulative_dose_response_pred_helper}
\title{function to calculate cumulative response}
\usage{
envpredutils.cumulative_dose_response_pred_helper(
  env_data_vector,
  .response_function.,
  parameters
)
}
\arguments{
\item{env_data_vector}{environmental covariate to predict on}

\item{.response_function.}{dose response function}

\item{parameters}{parameters of the response function}
}
\value{
returns vector with cumulative sum
}
\description{
wrapper function to apply the response_prediction
function in a "weighted" manner and do some gap filling of the env data
}
\keyword{internal}
