% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/DyMEP_apply_unknown_environment_prediction_utils.R
\name{envpredutils.data_frame_to_list}
\alias{envpredutils.data_frame_to_list}
\title{A function to get the best model with the covariates at hand}
\usage{
envpredutils.data_frame_to_list(env_data_frame)
}
\arguments{
\item{env_data_frame}{data.frame with the necessary environmental data,
one column must be "DATE" (as.Date format), the others with the names
 of the environmental covariates (e.g. tas, tasmin etc.)}
}
\description{
A function to get the best model with the covariates at hand
}
\examples{
envpredutils.data_frame_to_list(data.frame("tas"=c(1:100),
 "VPD"=c(1:100), "DATE"=c(1:100)))
}
\keyword{internal}
