% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/DyMEP_apply_unknown_environment_prediction_utils.R
\name{envpredutils.GLM_prediction_df_creator}
\alias{envpredutils.GLM_prediction_df_creator}
\title{create DF suitable to predict with GLM}
\usage{
envpredutils.GLM_prediction_df_creator(resposne_predictions, timestamp_vect)
}
\arguments{
\item{resposne_predictions}{prediction output}

\item{timestamp_vect}{vector with timestamps}
}
\value{
data.frame for glm prediction
}
\description{
create DF suitable to predict with GLM
}
\examples{
envpredutils.GLM_prediction_df_creator(list("tas"=list("growth_cumulative"=
                                    c(1))),timestamp_vect = Sys.Date())
}
\keyword{internal}
