% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/DyMEP_apply_unknown_environment_prediction_utils.R
\name{envpredutils.model_selecter}
\alias{envpredutils.model_selecter}
\title{A function to get the wanted model}
\usage{
envpredutils.model_selecter(
  crop_abbrev,
  pheno_phase,
  env_variables,
  external_params_path = NULL
)
}
\arguments{
\item{crop_abbrev}{abbreviation of the crop}

\item{pheno_phase}{phenological phase}

\item{env_variables}{vector with the wanted environmental variables}

\item{external_params_path}{path where additional crop parameters should be
stored if not possible to download in to the regular R repository. The
default is NULL, which will use the regular R repository as path}
}
\value{
model name from pre fitted models
}
\description{
A function to get the wanted model
}
\examples{
envpredutils.model_selecter("WW","sowing-emergence",c("tas"))
}
\keyword{internal}
