% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/DyMEP_apply_unknown_environment_prediction_utils.R
\name{envpredutils.period_date_creator}
\alias{envpredutils.period_date_creator}
\title{A function to create a date vector}
\usage{
envpredutils.period_date_creator(start_date, period_length)
}
\arguments{
\item{start_date}{start of the phase}

\item{period_length}{length of the period}
}
\description{
A function to create a date vector
}
\examples{
envpredutils.period_date_creator(as.Date("2024-01-15"),100)
}
\keyword{internal}
