% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/DyMEP_apply_unknown_environment_prediction_utils.R
\name{envpredutils.pheno_phase_prediction_glm_model}
\alias{envpredutils.pheno_phase_prediction_glm_model}
\title{apply the prediction with glm model}
\usage{
envpredutils.pheno_phase_prediction_glm_model(
  env_data_pheno_phase,
  pheno_phase,
  crop_abbrev,
  model,
  output_type = "dates"
)
}
\arguments{
\item{env_data_pheno_phase}{environmental data required to predict the phase}

\item{pheno_phase}{phenological phase}

\item{crop_abbrev}{abbreviation of the crop}

\item{model}{the selected model to predict the wanted phenological phase}

\item{output_type}{either "dates" or "detailed_information"; defines what
output of the model they user wants to have as return,
default is set to "dates". If a user wants to get the response parameters,
curves, predictions and model thresholds, it should be chosen
"detailed_information"
output = "dates" will return a dataframe with the stages and according dates
output = "detailed_information" will return a list with the dates, but also
 the corresponding dose response parameters and predictions}
}
\description{
apply the prediction with glm model
}
\keyword{glm}
\keyword{phase}
\keyword{prediction}
