% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyMEP_utils_prediction_helper_functions.R
\name{get_parameters}
\alias{get_parameters}
\title{A function to download additional crop parameters}
\usage{
get_parameters(external_params_path = NULL)
}
\arguments{
\item{external_params_path}{path where additional crop parameters should be
stored if not possible to download in to the regular R repository. The
default is NULL, which will use the regular R repository as path}
}
\value{
no return
}
\description{
A function to download additional crop parameters
}
\keyword{internal}
