% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyMEP_dose_response_curves.R
\name{non_linear_prediction}
\alias{non_linear_prediction}
\title{DRC function: non_linear_prediction}
\usage{
non_linear_prediction(env_variate, params)
}
\arguments{
\item{env_variate}{value of a environmental covariate}

\item{params}{list of input parameter; base_value: minimal value;
slope estimated value, slope of the linear phase}
}
\description{
broken stick model according to an env variable
}
\examples{
non_linear_prediction(1,list("base_value"=5,"slope_value"=1))

# visualization
non_linear <- lapply(seq(0, 10, 0.1),
non_linear_prediction,
list("base_value"=5,"slope_value"=1))
plot(seq(0, 10, 0.1), non_linear)
}
\keyword{DRC}
