% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyMEP_prediction_helper_functions.R
\name{predhelp.check_pheno_phase_order}
\alias{predhelp.check_pheno_phase_order}
\title{predhelp.check_pheno_phase_order}
\usage{
predhelp.check_pheno_phase_order(phase_covariate_list, crop_abbrev)
}
\arguments{
\item{phase_covariate_list}{list with all available phenology phases}

\item{crop_abbrev}{current crop_abbrev (name of the crop)}
}
\value{
returns the given phase_covariate_list but with correctly ordered
phenology phases
}
\description{
predhelp.check_pheno_phase_order
}
\section{Fields}{

\describe{
\item{\code{abbrev}}{Description of crop_abbrev column.}
}}

\examples{
predhelp.check_pheno_phase_order(phase_covariate_list = list(
  "sowing-emergence" = c("tasmin","VPD","SPI")),
  crop_abbrev = "WW")
}
\keyword{internal}
