% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyMEP_dose_response_curves.R
\name{reg_linear_prediction}
\alias{reg_linear_prediction}
\title{DRC function: reg_linear_prediction}
\usage{
reg_linear_prediction(env_variate, params)
}
\arguments{
\item{env_variate}{value of a environmental covariate}

\item{params}{list of input parameter; intercept estimated value,
slope of the linear phase}
}
\value{
value with reg_linear response
}
\description{
linear model according to an env variable
}
\examples{
reg_linear_prediction(1,list("intercept_value"=1, "slope_value"=5))
# visualization
reg_linear <- lapply(seq(0, 10, 0.1),
reg_linear_prediction,
list("intercept_value"=-1,"slope_value"=1))
plot(seq(0, 10, 0.1), reg_linear)
}
\keyword{DRC}
