% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyMEP_base_functions.R
\name{summary}
\alias{summary}
\title{Summary method for DyMEP class}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{An object of class "DyMEP".}

\item{...}{Additional arguments to be passed to summary functions.}
}
\description{
This function summarizes DyMEP objects.
}
\keyword{internal}
