% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/DyMEP_apply_unknown_environment_prediction_utils.R
\name{envpredutils.model_selecter}
\alias{envpredutils.model_selecter}
\title{A function to get the wanted model}
\usage{
envpredutils.model_selecter(crop_abbrev, pheno_phase, env_variables)
}
\arguments{
\item{crop_abbrev}{abbreviation of the crop}

\item{pheno_phase}{phenological phase}

\item{env_variables}{vector with the wanted environmental variables}
}
\description{
A function to get the wanted model
}
\examples{
envpredutils.model_selecter("WW","sowing-emergence",c("tas"))
}
\keyword{internal}
